pyproject.toml
[project]
name = "ap2-x402-demo"
version = "0.1.0"
description = "AP2 + x402 v2 + A2A demo — Merchant Agent with crypto payment via x402 protocol"
requires-python = ">=3.13"
dependencies = [
    "a2a-sdk",
    "google-adk>=0.5.0",
    "x402>=0.2.0",
    "click>=8.2.0",
    "eth-account>=0.13.7",
    "httpx>=0.28.1",
    "uvicorn>=0.34.2",
    "web3>=7.10.0",
    "flask>=3.1.0",
    "python-dotenv>=1.1.0",
    "x402-a2a",
]

[tool.uv.sources]
x402-a2a = { path = "../a2a-x402-ref/python/x402_a2a", editable = true }

[tool.hatch.build.targets.wheel]
packages = ["shared", "wallet", "merchant", "client"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
